backend/tests/unit/utils/test_db_operations.py
import pytest
from unittest.mock import Mock, patch
from backend.utils.db_operations import save_to_database

@pytest.fixture
def mock_db():
    with patch('your_module.db') as mock:
        yield mock

@pytest.fixture
def mock_logging():
    with patch('your_module.logging') as mock:
        yield mock

def create_bill_data():
    # Create a well-formed bill_data dictionary here
    return {
        'number': '123',
        'update_date': '2023-09-12',
        # ... (other keys)
    }

def test_save_to_database_with_complete_data(mock_db, mock_logging):
    bill_data = create_bill_data()
    
    save_to_database(bill_data)
    
    # Make assertions here, e.g.,
    # assert mock_db.session.add.call_count == expected_number_of_calls
    # assert mock_db.session.commit.call_count == expected_number_of_calls

def test_save_to_database_with_missing_data(mock_db, mock_logging):
    # Create a bill_data dictionary with missing keys here
    bill_data = create_bill_data()
    del bill_data['number']
    
    save_to_database(bill_data)
    
    # Make assertions here

def test_save_to_database_with_erroneous_data(mock_db, mock_logging):
    # Create a bill_data dictionary with erroneous data here
    bill_data = create_bill_data()
    bill_data['number'] = 'invalid_number'
    
    save_to_database(bill_data)
    
    # Make assertions here
